Trainer module has been added & Code Restructure
