+++ trainer/re-run-trainer.bat
@@ -1,0 +1,1 @@
+python -m spacy train config.cfg --output ..\output --paths.train pyresumize.spacy --paths.dev pyresumize.spacy
