+++ trainer/training_input.json
@@ -1,0 +1,10 @@
+[
+    {
+        "keyword": {
+            "string":"Gokul",
+            "start_at": 0,
+            "end_at":5,
+            "type":"PERSON"
+        }
+    }
+]
